pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "anthropic>=0.50.0",
    "fastapi>=0.115.12",
    "google-cloud>=0.34.0",
    "google-cloud-aiplatform>=1.90.0",
    "google-cloud-logging>=3.12.1",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.12.1",
    "mcp==1.8.1",
    "motor>=3.7.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
]
